Backend/requirements.txt
streamlit==1.44.1
fastapi==0.115.12
pydantic==2.11.3
uvicorn==0.34.0
mysql-connector-python==9.2.0
pandas==2.2.3
requests==2.32.3
pytest==8.3.5